Initial commit of quip script commands for Raycast
Two scripts for now:
- quip-new-note.py to create a new general note in Quip
- quip-new-meeting-note.py to create a new meeting note in Quip

Both scripts do the same, but store the notes in different folders in
Quip.

To use, add the quip folder as a script commands folder in Raycast. The
scripts will be identified and loaded automatically. You can then assign
aliases or hotkeys to them in the Raycast configuration.

The configuration of Quip API token, where to store the different types
of documents, etc., is configured in quip_config.py. This file is not
provided, rather quip_config.template.py, which must be configured and
renamed before the script runs.

The quip.py file contains the Python API library provided by Quip at
https://github.com/quip/quip-api/tree/master/python.

The quip_utils.py contains common functionality for creating and
managing Quip documents.