New configuration mechanism
Configuration is now done in quip_config.ini, an INI-style config file.

This file contains a DEFAULT section for common values, and defines the
document types that can be created in individual sections. Each section
can override any of the values and can set also a FolderID in which the
documents of that type will be stored.