New "Action" attribute
The Action attribute can be used to specify whether a new document will
be created (default value 'create'), or whether the given text should be
added to an existing document (value 'add'). The 'add' action requires
attribute 'DocID' to specify the document to which the text should be
added, and optionally 'ListType', which can be 'todo', 'bullet' or
'num', to specify different types of lists to use (by default plain text
is added at the end of the document). The text can contain markup in
Markdown format.